% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{annotateSE}
\alias{annotateSE}
\title{Construct summarized experiment}
\usage{
annotateSE(se, gtf)
}
\arguments{
\item{se}{a summarized experiment}

\item{gtf}{annotation}
}
\value{
a summarized experiment
}
\description{
Construct summarized experiment
}
