% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{computeRecovery}
\alias{computeRecovery}
\title{Computes abundance recovery rate}
\usage{
computeRecovery(
  se,
  sample,
  reference,
  log = TRUE,
  pseudocount = 1,
  statistic = "median",
  features = NULL
)
}
\arguments{
\item{se}{summarized experiment}

\item{sample}{samples to calculate ARR for}

\item{reference}{references (corresponding to samples)}

\item{log}{TRUE if data should be logged}

\item{pseudocount}{pseudocount for log}

\item{statistic}{statistic to aggregate by: mean or median}

\item{features}{features to calculate metric for}
}
\value{
a list object with plots and metrics
}
\description{
Computes abundance recovery rate
}
