% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{computeResEntropy}
\alias{computeResEntropy}
\title{Computes resolution entropy}
\usage{
computeResEntropy(
  se,
  samples,
  num.bins = seq(5, 20, length.out = 16),
  statistic = "median",
  features = NULL,
  log = TRUE,
  pseudocount = 1,
  pt.alpha = 0.5,
  pt.size = 2
)
}
\arguments{
\item{se}{summarized experiment}

\item{samples}{samples to calculate reproducibility}

\item{num.bins}{bins to test resolution entropy for}

\item{statistic}{statistic to aggregate metrics by}

\item{features}{features to calculate metric for}

\item{log}{TRUE if data should be logged}

\item{pseudocount}{pseudocount for log}

\item{pt.alpha}{alpha for points}

\item{pt.size}{size for points}
}
\value{
a list object with plots and metrics
}
\description{
Computes resolution entropy
}
