% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{constructSE}
\alias{constructSE}
\title{Construct summarized experiment}
\usage{
constructSE(path, methods)
}
\arguments{
\item{path}{path to quantification results}

\item{methods}{methods of corresponding results}
}
\value{
a summarized experiment
}
\description{
Construct summarized experiment
}
