% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Plot histogram}
\usage{
plotHistogram(
  se,
  samples,
  log = TRUE,
  pseudocount = 0.1,
  histogram = TRUE,
  density = FALSE,
  alpha = 0.5
)
}
\arguments{
\item{se}{summarized experiment}

\item{samples}{samples to calculate reproducibility}

\item{log}{TRUE if data should be logged}

\item{pseudocount}{pseudocount for log}

\item{histogram}{TRUE if histogram should be displayed}

\item{density}{TRUE if density should be displayed}

\item{alpha}{alpha for density}
}
\value{
histogram
}
\description{
Plot histogram
}
