% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotScatter}
\alias{plotScatter}
\title{Plot scatter}
\usage{
plotScatter(
  se,
  sample.x,
  sample.y,
  features = NULL,
  annotate = NULL,
  cor.method = "spearman",
  log = TRUE,
  pseudocount = 1,
  contour = TRUE,
  density = FALSE,
  diagonal = TRUE,
  pt.alpha = 0.5,
  pt.size = 3,
  contour.alpha = 0.5,
  contour.size = 0.7,
  density.alpha = 0.5
)
}
\arguments{
\item{se}{summarized experiment}

\item{sample.x}{sample 1}

\item{sample.y}{sample 2}

\item{features}{features to plot}

\item{annotate}{metadata to annotate by}

\item{cor.method}{correlation method}

\item{log}{TRUE if data should be logged}

\item{pseudocount}{pseudocount for log}

\item{contour}{TRUE if contours should be displayed}

\item{density}{TRUE if density should be displayed}

\item{diagonal}{TRUE if diagonal should be displayed}

\item{pt.alpha}{alpha for points}

\item{pt.size}{size for points}

\item{contour.alpha}{alpha for contour}

\item{contour.size}{size for contour}

\item{density.alpha}{alpha for density}
}
\value{
scatter plots
}
\description{
Plot scatter
}
